% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_to_amp.R
\name{cn_to_amp}
\alias{cn_to_amp}
\title{Converts continuous cn values to binary amp or not based on threshold}
\usage{
cn_to_amp(cn_df, threshold)
}
\arguments{
\item{cn_df}{a data frame of copy number values across samples}

\item{threshold}{a numeric value, the threshold over which you consider something to be amplified}
}
\description{
Takes a vector of cn values and a threshold, determines if values are over threshold and thus if amplification is present
}
