% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorByFeature.R
\name{colorByFeatureMain}
\alias{colorByFeatureMain}
\title{Map numeric values to a blue-to-red color gradient}
\usage{
colorByFeatureMain(vector)
}
\arguments{
\item{vector}{a vector of numeric values with sample names}
}
\description{
Takes a vector of values and returns an analogous vector of colors from a heatmap gradient of the vector values
}
