% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasets.R
\name{getDatasets}
\alias{getDatasets}
\title{Creates a dataframe of available data for each patient}
\usage{
getDatasets(datasets_path)
}
\arguments{
\item{datasets_path}{path to where datasets are located}
}
\description{
Looks through all files in the dataset path and creates a dataframe containing the datasets that exist for each tumor
}
