% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataValues.R
\name{getDataValues}
\alias{getDataValues}
\title{Retrieve the values for a particular patients dataset}
\usage{
getDataValues(
  patient,
  tumor,
  dataset,
  type,
  gene,
  threshold,
  conversion,
  tumorDatasetsPath
)
}
\arguments{
\item{patient}{Patient ID in long form (i.e. Patient300 not P300)}

\item{tumor}{sf number with sf lower}

\item{dataset}{name of dataset (i.e. Histology, Amplification, Copy Number, etc)}

\item{type}{only required if dataset is Histology}

\item{gene}{only required if dataset is Copy Number or RNAseq}

\item{threshold}{only required if dataset is Amplificaiton}

\item{conversion}{converts dataset to name of data file, where value is dataset and name is file name (ex c(cn.rds='Copy Number', ))}

\item{tumorDatasetsPath}{path to datasets}
}
\description{
Creates a vector of final data values for a particular patient, tumor, and dataset
}
