% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataValuesDownload.R
\name{getDataValuesDownload}
\alias{getDataValuesDownload}
\title{Retrieve the values for a particular patients dataset}
\usage{
getDataValuesDownload(
  patientsFinal,
  sfConversion,
  datasetD,
  typeD,
  rowSelectionD,
  thresholdD,
  conversion,
  tumorDatasetsPath,
  sampleData
)
}
\arguments{
\item{patientsFinal}{Patient ID in long form (i.e. Patient300 not P300)}

\item{sfConversion}{Coverts patientID to sf# in a one-to-one relationship}

\item{datasetD}{name of dataset (i.e. Histology, Amplification, Copy Number, etc)}

\item{typeD}{only required if dataset is Histology}

\item{rowSelectionD}{only required if dataset is Copy Number, RNAseq, Cell Types, Cancer-associated Processes, or Expansions}

\item{thresholdD}{only required if dataset is Amplification}

\item{conversion}{converts dataset to name of data file, where value is dataset and name is file name (ex c(cn.rds='Copy Number', ))}

\item{tumorDatasetsPath}{path to datasets}

\item{sampleData}{contains spatial information about each sample}
}
\description{
Creates a vector of final data values for a particular patient, tumor, and dataset
}
