% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3Dmodel.R
\name{plot3Dmodel}
\alias{plot3Dmodel}
\title{Plots 3D model}
\usage{
plot3Dmodel(patientID, sf, colors, tumorModelsPath)
}
\arguments{
\item{patientID}{Patient ID in long form (i.e. Patient300 not P300)}

\item{sf}{sf number with sf lower}

\item{colors}{vector of colors, usually corresponding to intensity of metric of interest; should be lenght of number of samples}

\item{tumorModelsPath}{path to models}
}
\value{
dolphins
}
\description{
Given a patientID, sf#, and colors, plots model of tumor with samples
}
\examples{
\donttest{\dontrun{plot3Dmodel('Patient300','sf10711',rep('#FF0000',10))}}
}
