% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3dSamples.R
\name{plot3DSamples}
\alias{plot3DSamples}
\title{Plots samples within a 3D model}
\usage{
plot3DSamples(sampleCoordinates, colors)
}
\arguments{
\item{sampleCoordinates}{Coordinates of samples with columns x y and z, each row is a sample with a row name}

\item{colors}{vector of colors, usually corresponding to intensity of metric of interest; should be length of number of samples and also have names}
}
\description{
Given a vector of named sample coordinates and colors, first determines if any samples are missing color info, then plots samples in an existing 3D model
}
