% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTemplate.R
\name{plotTemplate}
\alias{plotTemplate}
\title{Plots background tumor and brain template}
\usage{
plotTemplate(tumorModel, brainModel)
}
\arguments{
\item{tumorModel}{xyz binary mask originally derived from dicoms of tumor}

\item{brainModel}{xyz binary mask originally derived from dicoms of brain}
}
\description{
Given xyz binary masks of the tumor and brain, plots the background 3d model on which samples can later be added
}
