% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{launches the GLiomaAtlas3D app}
\usage{
launchApp()
}
\value{
shiny application object

\donttest{@examples launchApp()}
}
\description{
launches the GLiomaAtlas3D app
}
