% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3DModel.R
\name{plot3DModel}
\alias{plot3DModel}
\title{Plots 3D model}
\usage{
plot3DModel(
  patientID,
  sf,
  colors,
  tumorModelsPath = system.file(package = "GliomaAtlas3D", "exdata", "models")
)
}
\arguments{
\item{patientID}{Patient ID in long form (i.e. Patient300 not P300)}

\item{sf}{sf number with sf lower}

\item{colors}{vector of colors, usually corresponding to intensity of metric of interest; should be length of number of samples}

\item{tumorModelsPath}{path to models}
}
\description{
Given a patientID, sf#, and colors, plots model of tumor with samples
}
\examples{
plot3DModel("Patient300", sf="sf10711", colors=rep("#FF0000", times=10))
}
